+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 - `br`: added retained taxes CSLL, INSS and IRRF
+- `tax`: added support for `informative` tax categories that will be calculated and reported but will not affect the invoice totals.
+- `br`: made ISS an informative tax
 
 ## [v0.220.6] - 2025-08-12